@@ -1,8 +1,9 @@
+import prettierConfig from '@nnecec/prettier-config'
 /**
  * @type {import('prettier').Config&import('prettier-plugin-zh').ZhOptions}
  */
-module.exports = {
-  ...require('@nnecec/prettier-config'),
+export default {
+  ...prettierConfig,
   plugins: ['prettier-plugin-zh'],
   spaceAroundCode: true,
   spaceAroundLink: true,
