@@ -13,11 +13,7 @@ export default function HooksPage() {
         subtitle="A simple hook to motion element with X or Y."
         title="useOffsetMotion"
       />
-      <ArrowLink
-        href="/curation/hooks/use-interval"
-        subtitle="A countdown button with motion."
-        title="useInterval"
-      />
+      <ArrowLink href="/curation/hooks/use-interval" subtitle="A countdown button with motion." title="useInterval" />
     </div>
   )
 }
