@@ -13,11 +13,14 @@ export default function UseTimer() {
 
   return (
     <div>
-      <div>Click start button, it&apos;ll be finished after 3 seconds: <Chip>{str}</Chip></div>
-      <div>is running: <Chip>{`${isRunning}`}</Chip></div>
+      <div>
+        Click start button, it&apos;ll be finished after 3 seconds: <Chip>{str}</Chip>
+      </div>
+      <div>
+        is running: <Chip>{`${isRunning}`}</Chip>
+      </div>
       <div className="flex gap-2">
         <Button
-
           onClick={() => {
             setStr('pending.')
             start()
@@ -25,11 +28,8 @@ export default function UseTimer() {
         >
           start timer
         </Button>
-        <Button onClick={pause}>
-          pause timer
-        </Button>
+        <Button onClick={pause}>pause timer</Button>
         <Button
-
           onClick={() => {
             setStr('pending.')
             restart()
