@@ -9,11 +9,7 @@ export const Logo = () => {
   const onHoverEnd = () => setShow(false)
 
   return (
-    <div
-      className="flex cursor-pointer items-center"
-      onMouseEnter={onHoverStart}
-      onMouseLeave={onHoverEnd}
-    >
+    <div className="flex cursor-pointer items-center" onMouseEnter={onHoverStart} onMouseLeave={onHoverEnd}>
       <motion.div
         animate={show ? 'show' : 'hide'}
         className="h-[30px] w-[30px] bg-current"
@@ -22,7 +18,7 @@ export const Logo = () => {
           hide: { borderRadius: 0 },
           show: { borderRadius: '50%' },
         }}
-       />
+      />
     </div>
   )
 }
