@@ -1,15 +1,9 @@
-type Falsy = false | 0 | '' | null | undefined
+type Falsy = '' | 0 | false | null | undefined
 
 interface Array<T> {
-  filter<S extends T>(
-    predicate: BooleanConstructor,
-    thisArg?: any
-  ): Exclude<S, Falsy>[]
+  filter<S extends T>(predicate: BooleanConstructor, thisArg?: any): Exclude<S, Falsy>[]
 }
 
 interface ReadonlyArray<T> {
-  filter<S extends T>(
-    predicate: BooleanConstructor,
-    thisArg?: any
-  ): Exclude<S, Falsy>[]
+  filter<S extends T>(predicate: BooleanConstructor, thisArg?: any): Exclude<S, Falsy>[]
 }
