@@ -6,8 +6,7 @@ type Pos = {
 }
 
 // https://github.com/streamich/ts-easing/blob/master/src/index.ts
-const inOutCubic = (t: number) =>
-  t < 0.5 ? 4 * t * t * t : (t - 1) * (2 * t - 2) * (2 * t - 2) + 1
+const inOutCubic = (t: number) => (t < 0.5 ? 4 * t * t * t : (t - 1) * (2 * t - 2) * (2 * t - 2) + 1)
 
 function animate(callback: (progression: number) => void, duration = 3000, f = inOutCubic) {
   const start = performance.now()
@@ -34,8 +33,9 @@ export const useOffsetMotion = <T extends HTMLElement>(pos: Pos, duration?: numb
   const start = (f?: (t: number) => number) => {
     animate(
       progression => {
-        if (ref.current)
+        if (ref.current) {
           ref.current.style.transform = `translate3d(${x * progression}px, ${y * progression}px, 0)`
+        }
       },
       duration,
       f,
